+++ src/test/php/io/modelcontextprotocol/unittest/StdIoTest.class.php
@@ -1,7 +1,7 @@
 <?php namespace io\modelcontextprotocol\unittest;
 
-use io\modelcontextprotocol\{StdIo, CallFailed};
-use lang\Process;
+use io\modelcontextprotocol\StdIo;
+use lang\{Process, IllegalStateException};
 use test\{Assert, Expect, Test, Values};
 
 class StdIoTest {
@@ -12,7 +12,7 @@ class StdIoTest {
     $value= ['name' => 'test', 'version' => '1.0.0'];
     $fixture= new StdIo(new Process(PHP_BINARY, ['-r', "echo '{$this->result($value)}', PHP_EOL;"]));
 
-    Assert::equals($value, $fixture->call('test')->value());
+    Assert::equals($value, $fixture->call('test')->current()->value());
   }
 
   #[Test, Values([8192, 65536, 1048576])]
@@ -22,11 +22,11 @@ class StdIoTest {
       "echo json_encode(['jsonrpc' => '2.0', 'id' => '6100', 'result' => str_repeat('*', {$size})]), PHP_EOL;"
     ]));
 
-    Assert::equals($size, strlen($fixture->call('test')->value()));
+    Assert::equals($size, strlen($fixture->call('test')->current()->value()));
   }
 
-  #[Test, Expect(class: CallFailed::class, message: '#-1: Unexpected EOF from process')]
+  #[Test, Expect(class: IllegalStateException::class, message: 'Unexpected EOF from process')]
   public function unexpected_eof() {
-    (new StdIo(new Process(PHP_BINARY, ['-r', "exit(0);"])))->call('test');
+    (new StdIo(new Process(PHP_BINARY, ['-r', "exit(0);"])))->call('test')->next();
   }
 }